Refactor article components into a single index file
- Move ArticleFlatList, FeaturedArticleList, HorizontalCard, and FeaturedArticleCard into a single index.ts file
- Update import in protected index to use new component index
- Remove individual component files to simplify project structure
