+++ feature/article/components/display/cards/FeaturedCard.tsx
