+++ feature/article/components/display/cards/HorizontalCard.tsx
